% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/publishManifest.R
\docType{methods}
\name{publishManifest,PkgManifest,GistDest-method}
\alias{publishManifest,PkgManifest,GistDest-method}
\title{publish a manifest to Github in the form of a gist}
\usage{
\S4method{publishManifest}{PkgManifest,GistDest}(manifest, dest, desc, ctx, ...)
}
\arguments{
\item{desc}{description to apply to the Gist}

\item{ctx}{the github context to be used. By defualt this will open a Web
browser and ask you to authorize switchr to modify your gists.}

\item{...}{unused.}

\item{man}{The manifest to be published}
}
\value{
the url to access the raw
}
\description{
publish a manifest to Github in the form of a gist
}

