% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishManifest.R
\docType{methods}
\name{publishManifest,PkgManifest,GistDest-method}
\alias{publishManifest,PkgManifest,GistDest-method}
\alias{publishManifest,PkgManifest,GistDest}
\alias{publishManifest,SessionManifest,GistDest-method}
\alias{publishManifest,SessionManifest,GistDest}
\title{publishManifest}
\usage{
\S4method{publishManifest}{PkgManifest,GistDest}(manifest, dest,
  desc = "An R package manifest", fname = "manifest.rman", ...)

\S4method{publishManifest}{SessionManifest,GistDest}(manifest, dest,
  desc = "An R seeding manifest", fname = "manifest.rman", ...)
}
\arguments{
\item{manifest}{The manifest to be published}

\item{dest}{a GistDest object}

\item{desc}{description to apply to the Gist}

\item{fname}{The name of the file to create within the gist.}

\item{...}{unused}
}
\value{
the url to access the raw file within the gist.
}
\description{
publish a manifest to Github in the form of a gist
}
\examples{
man = libManifest()
\dontrun{gisturl = publishManifest(man, Gist())}
}
\references{
Becker G, Barr C, Gentleman R, Lawrence M; Enhancing Reproducibility and Collaboration via Management of R Package Cohorts. Journal of Statistical Software, 81(1). 2017. doi: 10.18637/jss.v082.i01
}
